feat(plugin): Remove FirebasePlugin nodes from AndroidManifest
In order to FirebaseIntercomPlugin to work properly, its configuration
must override the FirebasePlugin's on AndroidManifest.xml.

This commit adds a script that get rid of the FirebasePlugin's nodes on
the project's AndroidManifest.xml.

That allows, as a result, the FirebaseIntercomPlugin's configurations
to take place instead.

The script, that has been written in Typescript and compiled to
JavaScript, is executed on cordova's after-prepare hooks.

More: https://cordova.apache.org/docs/en/latest/guide/appdev/hooks/